+++ src/Form/Block/ContactFormType.php
@@ -1,0 +1,24 @@
+<?php
+/*
+ * (c) 2017: 975L <contact@975l.com>
+ * (c) 2017: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\ContactFormBundle\Form\Block;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class ContactFormType extends AbstractType
+{
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class'         => null,
+            'translation_domain' => 'contactForm',
+        ]);
+    }
+}